requirements.txt
python-binance
pandas
pandas-ta
requests
packaging
Flask
gunicorn